+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineEditorResources.cs
@@ -14,16 +14,18 @@ namespace UnityEngine.Rendering.HighDefinition
         public VolumeProfile defaultSkyAndFogProfile;
         [Reload("Editor/DefaultDXRScene/Sky and Fog Settings Profile.asset")]
         public VolumeProfile defaultDXRSkyAndFogProfile;
-        [Reload("Editor/DefaultScene/Scene PostProcess Profile.asset")]
-        public VolumeProfile defaultPostProcessingProfile;
-        [Reload("Editor/DefaultDXRScene/Scene PostProcess Profile.asset")]
-        public VolumeProfile defaultDXRPostProcessingProfile;
+        [Reload("Editor/DefaultDXRScene/DXR Settings.asset")]
+        public VolumeProfile defaultDXRSettings;
         [Reload(new[]
         {
             "Runtime/RenderPipelineResources/Skin Diffusion Profile.asset",
             "Runtime/RenderPipelineResources/Foliage Diffusion Profile.asset"
         })]
-        public DiffusionProfileSettings[] defaultDiffusionProfileSettingsList;
+        [SerializeField]
+        internal DiffusionProfileSettings[] defaultDiffusionProfileSettingsList;
+        
+        [Reload("Editor/RenderPipelineResources/DefaultSettingsVolumeProfile.asset")]
+        public VolumeProfile defaultSettingsVolumeProfile;
 
         [Serializable, ReloadGroup]
         public sealed class ShaderResources
@@ -80,9 +82,7 @@ namespace UnityEngine.Rendering.HighDefinition
         public ShaderGraphResources shaderGraphs;
         public LookDevResources lookDev;
     }
-
-
-
+    
     [UnityEditor.CustomEditor(typeof(HDRenderPipelineEditorResources))]
     class HDRenderPipelineEditorResourcesEditor : UnityEditor.Editor
     {