+++ com.unity.visualeffectgraph/Editor/Compiler/VFXGraphCompiledData.cs
@@ -38,6 +38,8 @@ namespace UnityEditor.VFX
 
     class VFXGraphCompiledData
     {
+        public const uint compiledVersion = 2;
+
         public VFXGraphCompiledData(VFXGraph graph)
         {
             if (graph == null)
@@ -53,6 +55,13 @@ namespace UnityEditor.VFX
             public VFXCompilationMode compilMode;
         }
 
+        private static VFXExpressionObjectValueContainerDesc<T> CreateObjectValueDesc<T>(VFXExpression exp, int expIndex)
+        {
+            var desc = new VFXExpressionObjectValueContainerDesc<T>();
+            desc.instanceID = exp.Get<int>();
+            return desc;
+        }
+
         private static VFXExpressionValueContainerDesc<T> CreateValueDesc<T>(VFXExpression exp, int expIndex)
         {
             var desc = new VFXExpressionValueContainerDesc<T>();
@@ -63,6 +72,11 @@ namespace UnityEditor.VFX
         private void SetValueDesc<T>(VFXExpressionValueContainerDesc desc, VFXExpression exp)
         {
             ((VFXExpressionValueContainerDesc<T>)desc).value = exp.Get<T>();
+        }
+
+        private void SetObjectValueDesc<T>(VFXExpressionValueContainerDesc desc, VFXExpression exp)
+        {
+            ((VFXExpressionObjectValueContainerDesc<T>)desc).instanceID = exp.Get<int>();
         }
 
         public uint FindReducedExpressionIndexFromSlotCPU(VFXSlot slot)
@@ -118,12 +132,12 @@ namespace UnityEditor.VFX
                         case VFXValueType.Texture3D:
                         case VFXValueType.TextureCube:
                         case VFXValueType.TextureCubeArray:
-                            value = CreateValueDesc<Texture>(exp, i);
+                            value = CreateObjectValueDesc<Texture>(exp, i);
                             break;
                         case VFXValueType.Matrix4x4: value = CreateValueDesc<Matrix4x4>(exp, i); break;
                         case VFXValueType.Curve: value = CreateValueDesc<AnimationCurve>(exp, i); break;
                         case VFXValueType.ColorGradient: value = CreateValueDesc<Gradient>(exp, i); break;
-                        case VFXValueType.Mesh: value = CreateValueDesc<Mesh>(exp, i); break;
+                        case VFXValueType.Mesh: value = CreateObjectValueDesc<Mesh>(exp, i); break;
                         case VFXValueType.Boolean: value = CreateValueDesc<bool>(exp, i); break;
                         default: throw new InvalidOperationException("Invalid type");
                     }
@@ -337,13 +351,13 @@ namespace UnityEditor.VFX
                                 {
                                     case VisualEffectAsset.PlayEventName:
                                         if( nextSg != null)
-                                            newEventPaths.Add(path.Concat(new int[] { 0 }).ToList());
+                                        newEventPaths.Add(path.Concat(new int[] { 0 }).ToList());
                                         else
                                             result[i].Add(evt);
                                         break;
                                     case VisualEffectAsset.StopEventName:
                                         if(nextSg != null)
-                                            newEventPaths.Add(path.Concat(new int[] { 1 }).ToList());
+                                        newEventPaths.Add(path.Concat(new int[] { 1 }).ToList());
                                         else
                                             result[i].Add(evt);
                                         break;
@@ -354,7 +368,7 @@ namespace UnityEditor.VFX
                                                 int eventIndex = nextSg.GetInputFlowIndex(eventName);
                                                 if(eventIndex != -1)
                                                     newEventPaths.Add(path.Concat(new int[] { eventIndex }).ToList());
-                                            }
+                                                }
                                             else
                                             {
                                                 result[i].Add(evt);
@@ -375,7 +389,7 @@ namespace UnityEditor.VFX
                                         newEventPaths.Add(path.Concat(new int[] { fixedSlotIndex }).ToList());
                                 }
                                 else
-                                    newEventPaths.Add(path.Concat(new int[] { currentFlowIndex }).ToList());
+                                newEventPaths.Add(path.Concat(new int[] { currentFlowIndex }).ToList());
                             }
                             else
                             {
@@ -657,11 +671,11 @@ namespace UnityEditor.VFX
 
             var eventDescTemp = new[]
             {
-                new { eventName = "OnPlay", playSystems = allPlayNotLinked, stopSystems = new List<uint>() },
-                new { eventName = "OnStop", playSystems = new List<uint>(), stopSystems = allStopNotLinked },
+                new { eventName = VisualEffectAsset.PlayEventName, playSystems = allPlayNotLinked, stopSystems = new List<uint>() },
+                new { eventName = VisualEffectAsset.StopEventName, playSystems = new List<uint>(), stopSystems = allStopNotLinked },
             }.ToList();
 
-            var specialNames = new HashSet<string>(new string[] {VisualEffectAsset.PlayEventName,VisualEffectAsset.StopEventName});
+            var specialNames = new HashSet<string>(new string[] {VisualEffectAsset.PlayEventName, VisualEffectAsset.StopEventName});
 
 
             var events = contexts.Where(o => o.contextType == VFXContextType.Event);
@@ -704,10 +718,11 @@ namespace UnityEditor.VFX
                 }
             }
             outEventDesc.Clear();
-            outEventDesc.AddRange(eventDescTemp.Select(o => new VFXEventDesc() { name = o.eventName, startSystems = o.playSystems.ToArray(), stopSystems = o.stopSystems.ToArray() }));
+            var relevantEvent = eventDescTemp.Where(o => o.playSystems.Any() || o.stopSystems.Any());
+            outEventDesc.AddRange(relevantEvent.Select(o => new VFXEventDesc() { name = o.eventName, startSystems = o.playSystems.ToArray(), stopSystems = o.stopSystems.ToArray() }));
         }
 
-        private static void GenerateShaders(List<GeneratedCodeData> outGeneratedCodeData, VFXExpressionGraph graph, IEnumerable<VFXContext> contexts, Dictionary<VFXContext, VFXContextCompiledData> contextToCompiledData, VFXCompilationMode compilationMode)
+        private void GenerateShaders(List<GeneratedCodeData> outGeneratedCodeData, VFXExpressionGraph graph, IEnumerable<VFXContext> contexts, Dictionary<VFXContext, VFXContextCompiledData> contextToCompiledData, VFXCompilationMode compilationMode, HashSet<string> dependencies)
         {
             Profiler.BeginSample("VFXEditor.GenerateShaders");
             try
@@ -725,7 +740,8 @@ namespace UnityEditor.VFX
 
                     if (context.doesGenerateShader)
                     {
-                        var generatedContent = VFXCodeGenerator.Build(context, compilationMode, contextData);
+                        
+                        var generatedContent = VFXCodeGenerator.Build(context, compilationMode, contextData, dependencies);
 
                         if(generatedContent!= null)
                         {
@@ -739,6 +755,9 @@ namespace UnityEditor.VFX
                         }
                     }
                 }
+
+                var resource = m_Graph.GetResource();
+
             }
             finally
             {
@@ -746,7 +765,7 @@ namespace UnityEditor.VFX
             }
         }
 
-        private static void SaveShaderFiles(VisualEffectResource resource, List<GeneratedCodeData> generatedCodeData, Dictionary<VFXContext, VFXContextCompiledData> contextToCompiledData)
+        private static VFXShaderSourceDesc[] SaveShaderFiles(VisualEffectResource resource, List<GeneratedCodeData> generatedCodeData, Dictionary<VFXContext, VFXContextCompiledData> contextToCompiledData)
         {
             Profiler.BeginSample("VFXEditor.SaveShaderFiles");
             try
@@ -767,8 +786,6 @@ namespace UnityEditor.VFX
                     descs[i].compute = generated.computeShader;
                 }
 
-                resource.shaderSources = descs;
-
                 for (int i = 0; i < generatedCodeData.Count; ++i)
                 {
                     var generated = generatedCodeData[i];
@@ -776,6 +793,8 @@ namespace UnityEditor.VFX
                     contextData.indexInShaderSource = i;
                     contextToCompiledData[generated.context] = contextData;
                 }
+
+                return descs;
             }
             finally
             {
@@ -919,6 +938,15 @@ namespace UnityEditor.VFX
                 var models = new HashSet<ScriptableObject>();
                 m_Graph.CollectDependencies(models,false);
 
+                var resource = m_Graph.GetResource();
+                resource.ClearSourceDependencies();
+
+                HashSet<string> sourceDependencies = new HashSet<string>();
+                foreach(VFXModel model in models.Where(t=> t is IVFXSlotContainer))
+                {
+                    model.GetSourceDependentAssets(sourceDependencies);
+                }
+
                 var contexts = models.OfType<VFXContext>().ToArray();
 
                 foreach (var c in contexts) // Unflag all contexts
@@ -1005,10 +1033,10 @@ namespace UnityEditor.VFX
                 var generatedCodeData = new List<GeneratedCodeData>();
 
                 EditorUtility.DisplayProgressBar(progressBarTitle, "Generating shaders", 7 / nbSteps);
-                GenerateShaders(generatedCodeData, m_ExpressionGraph, compilableContexts, contextToCompiledData, compilationMode);
+                GenerateShaders(generatedCodeData, m_ExpressionGraph, compilableContexts, contextToCompiledData, compilationMode, sourceDependencies);
 
                 EditorUtility.DisplayProgressBar(progressBarTitle, "Saving shaders", 8 / nbSteps);
-                SaveShaderFiles(m_Graph.visualEffectResource, generatedCodeData, contextToCompiledData);
+                VFXShaderSourceDesc[] shaderSources = SaveShaderFiles(m_Graph.visualEffectResource, generatedCodeData, contextToCompiledData);
 
                 var bufferDescs = new List<VFXGPUBufferDesc>();
                 var temporaryBufferDescs = new List<VFXTemporaryGPUBufferDesc>();
@@ -1050,9 +1078,9 @@ namespace UnityEditor.VFX
 
                 }
 
-                // Update renderer settings
-                VFXRendererSettings rendererSettings = GetRendererSettings(m_Graph.visualEffectResource.rendererSettings, compilableContexts.OfType<IVFXSubRenderer>());
-                m_Graph.visualEffectResource.rendererSettings = rendererSettings;
+                // Update transient renderer settings
+                ShadowCastingMode shadowCastingMode = compilableContexts.OfType<IVFXSubRenderer>().Any(r => r.hasShadowCasting) ? ShadowCastingMode.On : ShadowCastingMode.Off;
+                MotionVectorGenerationMode motionVectorGenerationMode = compilableContexts.OfType<IVFXSubRenderer>().Any(r => r.hasMotionVector) ? MotionVectorGenerationMode.Object : MotionVectorGenerationMode.Camera;
 
                 EditorUtility.DisplayProgressBar(progressBarTitle, "Setting up systems", 10 / nbSteps);
                 var expressionSheet = new VFXExpressionSheet();
@@ -1061,8 +1089,12 @@ namespace UnityEditor.VFX
                 expressionSheet.values = valueDescs.OrderBy(o => o.expressionIndex).ToArray();
                 expressionSheet.exposed = exposedParameterDescs.OrderBy(o => o.name).ToArray();
 
-                m_Graph.visualEffectResource.SetRuntimeData(expressionSheet, systemDescs.ToArray(), eventDescs.ToArray(), bufferDescs.ToArray(), cpuBufferDescs.ToArray(), temporaryBufferDescs.ToArray());
+
+                resource.SetRuntimeData(expressionSheet, systemDescs.ToArray(), eventDescs.ToArray(), bufferDescs.ToArray(), cpuBufferDescs.ToArray(), temporaryBufferDescs.ToArray(), shaderSources, shadowCastingMode, motionVectorGenerationMode,compiledVersion);
                 m_ExpressionValues = expressionSheet.values;
+
+                foreach (var dep in sourceDependencies)
+                    resource.AddSourceDependency(dep);
 
                 if (k_FnVFXResource_SetCompileInitialVariants != null)
                 {
@@ -1087,12 +1119,6 @@ namespace UnityEditor.VFX
             }
             finally
             {
-                EditorUtility.DisplayProgressBar(progressBarTitle, "Importing VFX", 11 / nbSteps);
-                Profiler.BeginSample("VFXEditor.CompileAsset:ImportAsset");
-                AssetDatabase.ImportAsset(assetPath, ImportAssetOptions.ForceUpdate); //This should compile the shaders on the C++ size
-                AssetDatabase.LoadAssetAtPath<VisualEffectAsset>(assetPath); // Force the reload to invoke the AwakeFromLoad as Import doesnt do it
-                Profiler.EndSample();
-
                 Profiler.EndSample();
                 EditorUtility.ClearProgressBar();
             }
@@ -1129,12 +1155,12 @@ namespace UnityEditor.VFX
                         case VFXValueType.Texture3D:
                         case VFXValueType.TextureCube:
                         case VFXValueType.TextureCubeArray:
-                            SetValueDesc<Texture>(desc, exp);
+                            SetObjectValueDesc<Texture>(desc, exp);
                             break;
                         case VFXValueType.Matrix4x4: SetValueDesc<Matrix4x4>(desc, exp); break;
                         case VFXValueType.Curve: SetValueDesc<AnimationCurve>(desc, exp); break;
                         case VFXValueType.ColorGradient: SetValueDesc<Gradient>(desc, exp); break;
-                        case VFXValueType.Mesh: SetValueDesc<Mesh>(desc, exp); break;
+                        case VFXValueType.Mesh: SetObjectValueDesc<Mesh>(desc, exp); break;
                         case VFXValueType.Boolean: SetValueDesc<bool>(desc, exp); break;
                         default: throw new InvalidOperationException("Invalid type");
                     }