+++ com.unity.visualeffectgraph/Editor/Models/Slots/Implementations/VFXSlotTexture3D.cs
@@ -7,11 +7,11 @@ using UnityEngine.Rendering;
 namespace UnityEditor.VFX
 {
     [VFXInfo(type = typeof(Texture3D))]
-    class VFXSlotTexture3D : VFXSlot
+    class VFXSlotTexture3D : VFXSlotObject
     {
         public override VFXValue DefaultExpression(VFXValue.Mode mode)
         {
-            return new VFXTexture3DValue(null, mode);
+            return new VFXTexture3DValue(0, mode);
         }
     }
 }